Add test

+++ CPUCode/TensorAddFile.cpp
@@ -162,7 +162,7 @@ void TensorAddFile::tensorAddLTensor() {
 	output.close();
 }
 
-const int loopSize =TensorAddition_tileSize;
+const long loopSize =TensorAddition_tileSize;
 
 void TensorAddFile::tensorAddDFE() {
 	long dim1, dim2, dim3, inTensorsLen =0, inTensorLen = 0;
@@ -202,12 +202,12 @@ void TensorAddFile::tensorAddDFE() {
 	}
 	//prepare data for DFE
 
-	int numZeroes = loopSize- inTensorLen % loopSize;
+	long numZeroes = loopSize- inTensorLen % loopSize;
 	inTensors = new float[inTensorsLen+numZeroes*numTensors];
-	int skipSize = loopSize*(numTensors-1);
-	int numLoaded=0;// counts to loopSize
-	int numLoaded1=0;//counts to tensor length
-	int offset=0;
+	long skipSize = loopSize*(numTensors-1);
+	long numLoaded=0;// counts to loopSize
+	long numLoaded1=0;//counts to tensor length
+	long offset=0;
 	long cnt=0;
 	float data;
 	for(long i=0;i<inTensorsLen;i++){
@@ -221,7 +221,7 @@ void TensorAddFile::tensorAddDFE() {
 			numLoaded=0;
 		}
 		if(numLoaded1==inTensorLen){
-			for(int j=0;j<numZeroes;j++){
+			for(long j=0;j<numZeroes;j++){
 				inTensors[cnt]=0;
 				cnt++;
 			}
@@ -243,7 +243,7 @@ void TensorAddFile::tensorAddDFE() {
 	clock_t end = clock();
 	double elapsed_time = (end - start)/(double)CLOCKS_PER_SEC;
 	cout << "Elapsed time dfe: " << elapsed_time << "\n";
-	for(int i=0;i<inTensorLen;i++){
+	for(long i=0;i<inTensorLen;i++){
 		output << sum[i] << " ";
 	}
 	delete[] inTensors;
